Se finaliza la estructura de los productos con su categoria y se finalizan las pruebas

+++ src/main/java/io/grisales/leonisa/ProntoPagable.java
@@ -1,5 +1,5 @@
 package io.grisales.leonisa;
 
 public interface ProntoPagable {
-    public double descProntoPago(int prontoPago);
+    public int descProntoPago(int prontoPago);
 }
